Refactor according to task
