+++ src/main/java/edu/hw2/task2/Square.java
@@ -1,6 +1,6 @@
 package edu.hw2.task2;
 
-public class Square extends BoundingRectangle {
+public class Square extends Rectangle {
     public Square(double sideLength) {
         super(sideLength, sideLength);
     }
